backend/requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
django==3.2
pymongo==3.12.1
djongo==1.3.6
sqlparse==0.2.4
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
psycopg2-binary==2.9.5
gunicorn==20.1.0
django-cors-headers==3.14.0
django-templated-mail==1.1.1
django-widget-tweaks==1.4.12
scikit-learn==1.2.2
numpy==1.24.2
google-cloud-storage
drf-spectacular==0.26.2